pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "flowstate_sdk"
version = "0.1.0"
description = "SDK for Agentic AI workflow management with Flowstate"
authors = [{name="Flowstate"}]
requires-python = ">=3.9"
license = { file = "LICENSE" }
dependencies = [
    "openai",
    "anthropic",
    "google-generativeai",
    "langchain-core",
    "python-dotenv"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
